# [RFC] Fix unnecessary busy waiting when no new tasks are incoming