@tsung-wei-huang 2022-10-28
@p12tic This is a great pull proposal! I will be happy to understand more details (yes I am aware of the issue you mentioned) by looking into the pull closely. Normally, I use [Taskflow Benchmarks](https://taskflow.github.io/taskflow/BenchmarkTaskflow.html) to analyze the performance difference between two versions. I have not yet had any good idea/time to automate this process like the usual regression. Any suggestion from you will be of great welcome :)