@p12tic 2022-10-29
@tsung-wei-huang Thanks for the suggestions.

As a preliminary test run I ran `ctest` on current master branch and on PR branch. The timings are as follows (this is on a workstation with AMD 2990WX with turbo boost enabled):

New (average 38.3s per run):
```
Total Test time (real) =  39.85 sec
Total Test time (real) =  37.95 sec
Total Test time (real) =  38.34 sec
Total Test time (real) =  38.06 sec
Total Test time (real) =  36.91 sec
Total Test time (real) =  37.08 sec
Total Test time (real) =  39.76 sec
Total Test time (real) =  37.30 sec
Total Test time (real) =  37.86 sec
Total Test time (real) =  40.05 sec
```

Old (overage 40.1s):
```
Total Test time (real) =  39.00 sec
Total Test time (real) =  40.37 sec
Total Test time (real) =  39.88 sec
Total Test time (real) =  39.36 sec
Total Test time (real) =  39.45 sec
Total Test time (real) =  41.30 sec
Total Test time (real) =  39.73 sec
Total Test time (real) =  41.38 sec
Total Test time (real) =  40.13 sec
Total Test time (real) =  40.74 sec
```

So there is around 4% performance improvement visible in unit tests already. I will run the proper benchmarks a bit later.

Regarding regression testing, to compare whether software version A is faster/slower than version B, a strategy of running the tests in ABABABABABAB order works even on unstable hardware that is dependent on environmental factors such as temperature. I've used this strategy successfully in the past to catch statistically significant performance regressions in mobile apps. It should be possible to run this on something like Github Actions even when we don't control the actual hardware the code gets to run on. This will not help tracking down performance differences over time, because there's large variation of performance parameters even on identical hardware and we don't control what runner the performance test will run on on Github side.