@tsung-wei-huang 2022-10-29
@p12tic thank you for the feedback. I am actually thinking if it is totally possible to remove the check:

```
 if(prev_num_actives == 0 && _num_actives) {
   _notifier.cancel_wait(worker._waiter);
   goto wait_for_task;
 }
```

because there is an empty check immediately following the four lines. Do you mind sharing some profile data again by removing the check?

However, I need to verify if this may cause bugs like missing tasks/wake-ups in some extreme situations.