@p12tic 2022-11-03
@tsung-wei-huang Thanks a lot for the question.

(This is a long comment, but I think you could find it interesting as there's small likelihood I've found a genuine improvement to the algorithm)

Turns out that your previous comment was spot on and I essentially broke the scheduling algorithm. The fact that there's at least one busy stealing worker is not a bug but an invariant of the design. I didn't previously read your latest paper about Taskflow in enough detail, now that I have read I see this as one of the core principles.

I have turned your example into a test and it verifies that the current PR breaks things. The test is included into new iteration of the PR.

Given that you know the library very well, I assume that the chance that I can improve it is pretty small, but I will still explore alternatives a bit, as one busy stealing worker is very non-optimal in my use case.

The objective is to park the currently busy stealing worker in cases when there's no new tasks. In order to do this we need to somehow wake at least one worker when there's a new task and there's no busy stealing worker.

Suppose we allow the last stealing worker to park once it doesn't find any tasks to steal. At that point the following holds:

 - There are zero stealing workers.

 - All workers have exactly zero tasks in their queues.

We can pick this situation as a synchronization point. When a worker schedules a new task, if it notices that its queue was empty just before adding a new task to it, it could check if there are zero stealing workers and wake at least one up. More formally:

    // last stealing worker just before parking
    num_thieves = 0
    if all queues empty:
        park()

    // worker that schedules a new task to its empty queue
    queue_was_empty = queue.empty()
    queue.push(task)
    if num_thieves == 0 && queue_was_empty:
        notify_worker()

It's evident `num_thieves = 0` in the parking worker must happen before `queue.push(task)` of any scheluling worker in order for the thread to park. As a consequence the scheduling worker is guarantee to observe `num_thieves == 0` in case the parking thread parked and will be able to notify it.

This PR now implements the approach outlined above. The last commit optimizes it a bit by extracting the information of whether `num_thieves == 0` is true into a separate variable. This makes the checks of `num_thieves == 0` essentially free as the value can live in private caches of the processing cores most of the time. This is not possible with num_thieves because its value changes often.

I have run benchmarks from the `benchmarks` directory and did not observe any change of performance. One note is that I run comparisons with "Return whether TaskQueue was empty before push()" commit not the master branch because at least GCC 10.2 optimizes the function slightly differently (verified with dissassembler) which introduces around 3% performance regression depending on benchmark. I will resolve this by a separate PR as it's pre-existing missed optimization.