@p12tic 2022-11-03
Just copying a comment I made on #442.

I've setup some counters and the frequency of `_notifier.notify()` calls from `_schedule()` are minuscule as implemented in #440. It's on the order of less than ten per second.