@p12tic 2022-11-03
@tsung-wei-huang The thing that makes it possible to have a very small number of `_notifier.notify(...)` calls from `_schedule(...)` is this commit https://github.com/taskflow/taskflow/pull/440/commits/9273952249a2395f594c34b90553d1f85dc17500. Basically, we ensure that `_notifier.notify(...)` is unlikely to be called more than once in cases when it's needed.