@tsung-wei-huang 2022-11-07
Hi @p12tic,

Sorry for the late response as I have been thinking of a better way to put everything together. Thank you for the pull request. It inspires me to rethink a new work-stealing algorithm to avoid busy stealing when no new tasks are coming. I summarize the solution below (might be a bit long to read):

**Recap**:  The busy waiting problem happened in the design of Taskflow's [scheduling algorithm](https://taskflow.github.io/papers/tpds21-taskflow.pdf). The algorithm maintains the invariant -- *when there is an active worker running tasks, we should keep at least a worker busy in stealing unless all are active*. This prevents starvation and keeps over-utilized threads upper-bounded by 1. Of course, in the extreme case where you have a long linear chain of tasks, the busy worker making stealing attempts become totally non-necessary.

**Motivation**: To solve the problem, I have designed a new invariant based on our discussion -- *when a task queue of a worker is non-empty, we should keep at least one worker busy in stealing unless all workers are busy in running tasks". This new invariant is different from the original one in the sense that no workers will be busy making stealing attempts if all queues are empty, in particular, when a worker is running a long task and its queue is empty. So, if we can keep this new invariant in our algorithm, we are good :)

**Solution**: I have pushed my solution to both the master and the dev branches. We can realize this new invariant by implementing the following logic:

  1. When a task queue of a worker becomes non-empty (i.e., from `0` to `1`) AND there is no thieves making stealing attempts, we should notify. [See code here.](https://github.com/taskflow/taskflow/blob/e654075e3104a1b7127c2772ef0bf6a720bfe110/taskflow/core/executor.hpp#L1230)
  2. When a task drains out its local task queue, it becomes a thief (`++_num_thieves`) and starts to make stealing attempts. There are two situations here: If the steal is successful and it is the last thief, it notifies another thief to avoid starvation. Otherwise, it moves on to the 2-phase committee guard, decrements the thieve count (`--_num_thieves`), and check if all queus are empty before committing to sleep. [See code here.](https://github.com/taskflow/taskflow/blob/e654075e3104a1b7127c2772ef0bf6a720bfe110/taskflow/core/executor.hpp#L1092)

**Proof Sketch**: Consider the following two situations:

1. A worker inserts a task to its *empty* queue: here, if `_num_thieves ==0`, we have the easy case to notify; otherwise, we know at least one thieves will check all queues again before committing to sleep (notice [here](https://github.com/taskflow/taskflow/blob/e654075e3104a1b7127c2772ef0bf6a720bfe110/taskflow/core/executor.hpp#L1113) we decrement the thieve count before this checking to resemble Dekker algorithm)  
2. A worker inserts a task to its *non-empty* queue: here, we will by-pass expensive calls to `_num_thieves` and `nofity`, but can we guarantee no starvation happens? Yes, by indication, the 1st proof will guarantee at least one thieve will eventually steal this worker's queue and check if it is empty. Furthermore,  [if the thieve is the last, it will wake up another one to steal tasks](https://github.com/taskflow/taskflow/blob/e654075e3104a1b7127c2772ef0bf6a720bfe110/taskflow/core/executor.hpp#L1104).

**Benefit** The new invariant largely simplifies the scheduling algorithm. We now remove `_num_actives` which was required for the original invariant. Also, the `_wait_for_tasks` is also simplified without any access to `_num_actives`. The cost of accessing `_num_thieves` is not much for a running worker because it only happens when that worker inserts a task into an empty queue. I am also seeing a boost of performance with this new invariant:

```bash
# original invariant (with busy stealing)
Total Test time (real) =  20.13 sec
Total Test time (real) =  19.94 sec
Total Test time (real) =  19.90 sec
Total Test time (real) =  19.48 sec
Total Test time (real) =  19.58 sec
Total Test time (real) =  19.86 sec
Total Test time (real) =  19.82 sec
Total Test time (real) =  20.10 sec
Total Test time (real) =  19.82 sec
Total Test time (real) =  19.83 sec

# new invariant (without busy stealing)
Total Test time (real) =  19.11 sec
Total Test time (real) =  18.75 sec
Total Test time (real) =  19.19 sec
Total Test time (real) =  18.83 sec
Total Test time (real) =  19.02 sec
Total Test time (real) =  18.97 sec
Total Test time (real) =  19.09 sec
Total Test time (real) =  18.96 sec
Total Test time (real) =  19.04 sec
Total Test time (real) =  19.05 sec
```

Also, I ran the following program and compared its CPU utilization between the two invariants:

```cpp
tf::Executor executor;
tf::Taskflow taskflow();
taskflow.emplace([]() { while(1); });  
executor.run(taskflow).wait();
```

Below is the CPU utilization using `top` command (`199.7` vs `100.3`):

```bash
# original invariant 
2993445 twhuang   20   0  596240   3428   3140 S 199.7   0.0   0:16.55 simple 

# new invariant
2159625 twhuang   20   0  596240   3376   3092 S 100.3   0.0   0:19.82 simple
```

**Action from you** If possible, could you please test the new version (main branch) on your workload again? The implementation does not have any additional atomic variables like `_last_thief_needs_wake_up ` you created in this pull. Also, I have verified the implementation is correct in terms of avoiding starvation and over-subscription of threading. Additionally, I will appreciate you modify the pull by migrating your unittest to the new implementation.

Thank you! This is a big contribution from you :) 

Probably we can consider submitting our result to [JSSPP](https://jsspp.org/).